Rename `OrderCreateCouponEdition` classes to `OrderCreateCouponEdit`
