+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/creation/coupon/OrderCreateCouponEditScreen.kt
@@ -18,12 +18,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.woocommerce.android.R
 import com.woocommerce.android.ui.compose.component.WCColoredButton
 import com.woocommerce.android.ui.compose.component.WCOutlinedTextField
-import com.woocommerce.android.ui.orders.creation.coupon.OrderCreateCouponEditionViewModel.ValidationState.ERROR
-import com.woocommerce.android.ui.orders.creation.coupon.OrderCreateCouponEditionViewModel.ValidationState.IDLE
+import com.woocommerce.android.ui.orders.creation.coupon.OrderCreateCouponEditViewModel.ValidationState.ERROR
+import com.woocommerce.android.ui.orders.creation.coupon.OrderCreateCouponEditViewModel.ValidationState.IDLE
 
 @Composable
-fun OrderCreateCouponEditionScreen(
-    state: State<OrderCreateCouponEditionViewModel.ViewState?>,
+fun OrderCreateCouponEditScreen(
+    state: State<OrderCreateCouponEditViewModel.ViewState?>,
     onCouponCodeChanged: (String) -> Unit,
     onCouponRemoved: () -> Unit,
 ) {
@@ -59,10 +59,10 @@ fun OrderCreateCouponEditionScreen(
 @Preview
 @Composable
 fun OrderCreateCouponEditionScreenPreview() {
-    OrderCreateCouponEditionScreen(
-        state = object : State<OrderCreateCouponEditionViewModel.ViewState?> {
-            override val value: OrderCreateCouponEditionViewModel.ViewState
-                get() = OrderCreateCouponEditionViewModel.ViewState(true, "code", true, IDLE)
+    OrderCreateCouponEditScreen(
+        state = object : State<OrderCreateCouponEditViewModel.ViewState?> {
+            override val value: OrderCreateCouponEditViewModel.ViewState
+                get() = OrderCreateCouponEditViewModel.ViewState(true, "code", true, IDLE)
         },
         onCouponCodeChanged = {},
         onCouponRemoved = {}
